Improve comment loading performance
